% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.affectgrid.R
\name{affect_grid}
\alias{affect_grid}
\title{Process affect-grid task data file}
\usage{
affect_grid(path, p)
}
\arguments{
\item{path}{Path to data file}

\item{p}{Participant identifier}
}
\value{
Data frame
}
\description{
\code{affect_grid()} processes data files generated by the affect-grid task
(\url{https://earcanal.github.io/affect-grid/}).
}
\keyword{affect}
\keyword{expfactory}
\keyword{grid}
