% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.affectgrid.R
\name{affect_grid}
\alias{affect_grid}
\title{Process affect-grid task data file}
\usage{
affect_grid(path, p)
}
\arguments{
\item{path}{Path to data file}

\item{p}{Participant identifier}
}
\value{
Data frame
}
\description{
\code{affect_grid()} processes data files generated by the affect-grid task
(\url{https://earcanal.github.io/affect-grid/}).
}
\examples{
\dontrun{
rsvp_df <- expand.grid(token = participants$token) \%>\%
 mutate(path=paste('data/', token, '_finished/rsvp-task-results.json', sep=''))
rsvp_df <- rsvp_df \%>\%
group_by(token) \%>\%
 do(., process_rsvp(.$path, .$token)) \%>\%
 ungroup \%>\%
 left_join(participants, by='token') \%>\%
 filter(phase == 'test', test_part == "response") \%>\%
 select(p, condition, trial_number, lag, correct, t1_correct, t2_correct, rt)
}
}
\keyword{affect}
\keyword{expfactory}
\keyword{grid}
